requirements.txt
aiogram
aiomysql
APScheduler
pytesseract
yoomoney
crystalpayio
requests
fastapi
cryptography
exceptiongroup==1.2.0
fastapi==0.110.0
frozenlist==1.4.1
idna==3.6
magic-filter==1.0.12
multidict==6.0.5
packaging==24.0
pillow==10.2.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyMySQL==1.1.0
pytesseract==0.3.10
pytz==2024.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
tzlocal==5.2
urllib3==2.2.1
yarl==1.9.4
YooMoney==0.1.0
